# api-fitness-foods-lc/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOC API",
        "description": "Fitness Foods LC API",
        "version": "1.0"
    },
    "paths": {
        "/api/errors": {
            "get": {
                "summary": "List all products",
                "operationId": "db7a3042150734bcb3f8d014d19dcdeb",
                "parameters": [
                    {
                        "name": "message",
                        "in": "query",
                        "description": "message of error"
                    },
                    {
                        "name": "message_exception",
                        "in": "query",
                        "description": "message_exception of error"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                }
            }
        },
        "/api/products/{code}": {
            "get": {
                "summary": "List product by code",
                "operationId": "38ef2921d58d1f82bcdcd02516bd488b",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code of products",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "summary": "Atualizar informações de um produto por código",
                "operationId": "321e73e3ec07008d6b2a675189400977",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "http://world-en.openfoodfacts.org/product/0000000000017/vitoria-crackers"
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Vitória crackers"
                                    },
                                    "quantity": {
                                        "type": "string",
                                        "example": "10"
                                    },
                                    "brands": {
                                        "type": "string",
                                        "example": "Vigor"
                                    },
                                    "categories": {
                                        "type": "string",
                                        "example": "Queijp"
                                    },
                                    "labels": {
                                        "type": "string",
                                        "example": "Requeijapo cremoso"
                                    },
                                    "cities": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "purchase_places": {
                                        "type": "string",
                                        "example": "Mercados"
                                    },
                                    "stores": {
                                        "type": "string",
                                        "example": "Pao de acucar"
                                    },
                                    "ingredients_text": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "traces": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "serving_size": {
                                        "type": "string",
                                        "example": "100g"
                                    },
                                    "serving_quantity": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "nutriscore_score": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "nutriscore_grade": {
                                        "type": "string",
                                        "example": "20"
                                    },
                                    "main_category": {
                                        "type": "string",
                                        "example": "Laticinios"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://static.openfoodfacts.org/images/products/000/000/000/0017/front_fr.4.400.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Delete product by code",
                "operationId": "f7bab3c82f77afd6b1e717140648abf9",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code of products",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado!"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "List all products",
                "operationId": "168342d1a92c4f2fd68833113724f9e5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page of result"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                }
            }
        }
    }
}